rename filename diffllama into diff_llama
